packet train

+++ net/ipv4/tcp_cubic.c
@@ -107,7 +107,7 @@ int round_id = 0;
 u32 last_round_start = 0;
 u64 last_ack_bytes_sent = 0;
 u32 last_packet_time = 0;
-u64 last_packet_bytes = 0;
+u32 last_packet_bytes = 0;
 
 static inline void bictcp_reset(struct bictcp *ca)
 {
@@ -426,18 +426,18 @@ static void hystart_update(struct sock *sk, u32 delay)
 			{
 				// u64 curr_bytes_sent = tp->bytes_sent - last_ack_bytes_sent;
 				// u64 bitrate = (curr_bytes_sent * 8 * 1000000 / (now - ca->round_start)) / (1000 * 1000);
-				u64 packet_bytes = tp->snd_una - last_packet_bytes;
+				u32 packet_bytes = tp->snd_una - last_packet_bytes;
 				u64 packet_time = now - last_packet_time;
 				if (packet_time > 0 && packet_time <= 2000)
 				{
 					u64 est_bd = (packet_bytes * 8 * 1000000 / packet_time) / (1000 * 1000);
-					printk(KERN_INFO "CUBIC (port: %hu) [Round %hu] Now %u, Round Start %u, Bytes ACKed %lld, Est. bandwidth %lld Mb/s\n",
+					printk(KERN_INFO "CUBIC (port: %hu) [Round %hu] Now %u, Round Start %u, Bytes ACKed %d, Est. bandwidth %d Mb/s\n",
 						port, round_id, now, ca->round_start, tp->snd_una, est_bd);
 				}
 				else if (packet_time > 2000)
 				{
 					u64 est_bd = (packet_bytes * 8 * 1000000 / packet_time) / (1000 * 1000);
-					printk(KERN_INFO "CUBIC (port: %hu) [Round %hu] Now %u, Round Start %u, Bytes ACKed %lld, Est. bandwidth %lld Mb/s\n",
+					printk(KERN_INFO "CUBIC (port: %hu) [Round %hu] Now %u, Round Start %u, Bytes ACKed %d, Est. bandwidth %d Mb/s\n",
 						port, round_id, now, ca->round_start, tp->snd_una, est_bd);
 					last_packet_bytes = tp->snd_una;
 					last_packet_time = now;
